finished the table as root layout optimization (but still needs to be tested)
